requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
numpy==1.21.2
Pillow==8.4.0
tensorflow==2.6.0
